+++ app.py
@@ -1,12 +1,11 @@
 import os
-import psycopg2
+import atexit
 from flask import Flask, jsonify
 from cfenv import AppEnv
-from psycopg2.extras import RealDictCursor
-from psycopg2.pool import ThreadedConnectionPool
-#from psycopg2.pool import SimpleConnectionPool
-#from psycopg2 import pool
 from flask_cors import CORS
+from psycopg_pool import ConnectionPool
+from psycopg.rows import dict_row
+from psycopg.errors import OperationalError
 
 origin = os.getenv("ORIGIN")
 port = int(os.getenv("PORT", 8080))
@@ -14,59 +13,48 @@ port = int(os.getenv("PORT", 8080))
 app = Flask(__name__)
 CORS(app, origins=origin)
 
-#app_env = AppEnv()
-#aws_rds = app_env.get_service(name="example-website-api-database")
+app_env = AppEnv()
+aws_rds = app_env.get_service(name="example-website-api-database")
 
-#Uncomment after testing
-"""
-connection_pool = ThreadedConnectionPool(url=database_url
-    minconn=5,
-    maxconn=20,
-    host=aws_rds.credentials.get("host"),
-    user=aws_rds.credentials.get("username"),
-    password=aws_rds.credentials.get("password"),
-    database=aws_rds.credentials.get("name"),
-    port=aws_rds.credentials.get("port"),
-)
-"""
-
-
-#Local Testing below
-#connection_pool uses psycopg2 under the hood
-
-connection_pool = ThreadedConnectionPool(
-    minconn=1,
-    maxconn=os.getenv('DB_MAX_CONNECTIONS'),
-    host=os.getenv('DB_HOST'),
-    user=os.getenv('DB_USER'),
-    password="pgpassword",
-    database="app_db",
-)
-
-# Retrieve connection from pool
-# implement try/catch 1. check if this connected 2. get connection and if fails with error then reestablish connection pool and retry within catch
-# ex. try return connection_pool.getconn() except -- and make sure it doesnt loop endlessly 
-
-def getconnection():
-    return connection_pool.getconn()
-
-#Returns connection to pool for reuse
-def returnconnection(db_connection, close=False):
-    connection_pool.putconn(db_connection, close=close)
-"""
-
-# Test Function
-def get_db_connection():
-    conn = psycopg2.connect(
-        host="postgres",
-        database="app_db",
-        user="pguser",
-        password="pgpassword"
+# Set up pool as a function so in the case it dies we can recreate it 
+def setup_connection_pool():
+    
+    database_url = os.getenv("DATABASE_URL")
+    
+    pool = ConnectionPool(
+        conninfo=database_url,
+        min_size=1,
+        max_size=20,
     )
-    conn.autocommit = True
-    return conn
+    return pool
 
-"""
+# Immediately create pool at module import 
+connection_pool = setup_connection_pool()
+
+# Maybe dont need atexit if using context manager per psycopg3 docs
+atexit.register(connection_pool.close)
+
+# Recreate pool if it dies
+def recreate_connection_pool():
+    global connection_pool
+    
+    # Close if it exists
+    if connection_pool is not None:
+        try:
+            connection_pool.close()
+        except Exception:
+            pass
+        
+    # Create new pool   
+    new_pool = setup_connection_pool()
+    
+    # Update variable
+    connection_pool = new_pool
+    
+    atexit.register(connection_pool.close)
+    
+    return connection_pool
+
 @app.route("/", methods=["GET"])
 def hello():
     try:
@@ -74,54 +62,34 @@ def hello():
     except Exception as error:
         return jsonify({"error": str(error)}), 500
 
-"""
 @app.route("/get_table", methods=["GET"])
-def get_table():
-    db_connection = None
+def get_table():   
+    global connection_pool
     try:
-        db_connection = getconnection()    
-        cursor = db_connection.cursor(cursor_factory=RealDictCursor)
-
-        query = "SELECT * FROM fdic_banks LIMIT 15"
-        cursor.execute(query)
-
-        rows = cursor.fetchall()
-        cursor.close()
-
-        # Specific to data from db to be converted to json
+        # Using the method and not the pool itself as the context manager
+        with connection_pool.connection() as conn:
+                cursor = conn.cursor(row_factory=dict_row)
+                cursor.execute("SELECT * FROM fdic_banks LIMIT 15")
+                rows = cursor.fetchall()
         return jsonify(rows)
-    except Exception as error:
-        # implement both error logs (app.logger) server way / dont stringify entire error to go back to user / replace 500 with "unexpected error occured"
-        return jsonify({"error": str(error)}), 500
-    finally: 
-        if db_connection:
-            returnconnection(db_connection)
-"""
 
-# Test query
-@app.route("/get_user", methods=["GET"])
-def get_user():
-    db_connection = None
+    except OperationalError:
+    #DB error, recreate pool
+        connection_pool = recreate_connection_pool()
+    
+    # Use new pool
     try:
-        db_connection = getconnection()
-        cursor = db_connection.cursor(cursor_factory=RealDictCursor)
-        
-        query = "SELECT * FROM users LIMIT 1"
-        cursor.execute(query)
-        
-        user = cursor.fetchone()
-        cursor.close()
-        
-        if user:
-            return jsonify(user)
-        else:
-            return jsonify({"message": "No users found"}), 404
-        
+        with connection_pool.connection() as conn:
+                cursor = conn.cursor(row_factory=dict_row)
+                cursor.execute("SELECT * FROM fdic_banks LIMIT 15")
+                rows = cursor.fetchall()
+        return jsonify(rows)
+    
+    except Exception as error:
+        return jsonify({"error": "Database connection failed"}), 500
+    
     except Exception as error:
         return jsonify({"error": str(error)}), 500
-    finally:
-        if db_connection:
-            returnconnection(db_connection)
-
+    
 if __name__ == "__main__":
     app.run(host="0.0.0.0", port=port, threaded=True)
